huge database rework

+++ app/src/main/java/com/filimonov/mycrochet/data/Project.kt
@@ -1,13 +1,16 @@
 package com.filimonov.mycrochet.data
 
+/**
+ * Represents project
+ */
 data class Project(
     val id: Int,
     val name: String,
     val description: String,
     val link: String,
-    val crochetSize: Int
+    val crochetSize: Float
 ) {
     companion object {
-        val Empty = Project(id = -1, name = "", description = "", link = "", crochetSize = 0)
+        val Empty = Project(id = -1, name = "", description = "", link = "", crochetSize = 0f)
     }
 }
