+++ app/src/main/java/com/filimonov/mycrochet/data/db/ProjectEntity.kt
@@ -9,5 +9,5 @@ data class ProjectEntity(
     val name: String,
     val description: String,
     val link: String,
-    val crochetSize: Int
+    val crochetSize: Float
 )
