+++ app/src/main/java/com/filimonov/mycrochet/ui/screens/projects/ProjectsViewModel.kt
@@ -13,7 +13,7 @@ import kotlinx.coroutines.launch
 class ProjectsViewModel(private val repository: ProjectsRepository) : ViewModel() {
     val projects: StateFlow<List<Project>> = repository.getProjects().stateIn(viewModelScope, started = SharingStarted.Lazily, initialValue = emptyList())
 
-    fun addProject(name: String, description: String, link: String, crochetSize: Int) {
+    fun addProject(name: String, description: String, link: String, crochetSize: Float) {
         viewModelScope.launch(Dispatchers.IO) {
             repository.addProject(
                 Project(id = 0, name = name, description = description, link = link, crochetSize = crochetSize)
